feat: move workspace/utils to core
