+++ src/core/public/utils/index.ts
@@ -31,10 +31,11 @@
 export { shareWeakReplay } from './share_weak_replay';
 export { Sha256 } from './crypto';
 export { MountWrapper, mountReactNode } from './mount';
-export { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId } from './workspace';
 export {
   WORKSPACE_PATH_PREFIX,
   PUBLIC_WORKSPACE_ID,
   MANAGEMENT_WORKSPACE_ID,
   WORKSPACE_TYPE,
+  formatUrlWithWorkspaceId,
+  getWorkspaceIdFromUrl,
 } from '../../utils';
