+++ src/core/server/utils/index.ts
@@ -32,3 +32,4 @@ export * from './crypto';
 export * from './from_root';
 export * from './package_json';
 export * from './streams';
+export { getWorkspaceIdFromUrl, cleanWorkspaceId } from '../../utils';
